feat: rename updateLbb controller
